Add files via upload
My Dad turned 60 this year and so to honour him from across 'the pond,' I made him a website full of his (exaggerated) accomplishments. I had just started coding and so this was my first project to practice the HTML5 and CSS3 that I had learnt so far. I photoshopped some images of him and our family and wrote him a poem about his 'life.' It was all very fun and he enjoyed it a lot - success!